Fix concurrent build calls
Two requests could build the same object in the app scope.
The first one was used by the first request, while the second one was
used by all the other requests. The first one was forgotten and never closed.
It could happen if the first two Get calls were done at the same time
or/and if the Build function took some time to be executed.