Fix the object closing order
The objects were closed in a random order. It was possible to have an
object with a closed dependency. It could lead to useless errors (and
error logs). In the worst case the dependency could be a logger and the
logs could be lost.

The objects are now closed in the right order. The dependencies are
closed after the objects using them.